rename(domain-jpa): Repository 구현체 이름 변경 (JpaAdapter->JpaRepository)
