+++ modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/device/DeviceEntityRepository.java
@@ -1,0 +1,9 @@
+package com.whoz_in.domain_jpa.device;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DeviceEntityRepository extends JpaRepository<DeviceEntity, Long> {
+
+}
