--- modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/device/DeviceJpaAdapter.java
@@ -1,16 +1,0 @@
-package com.whoz_in.domain_jpa.device;
-
-import com.whoz_in.domain.device.DeviceRepository;
-import com.whoz_in.domain.device.model.Device;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class DeviceJpaAdapter implements DeviceRepository {
-    private final DeviceJpaRepository repository;
-    @Override
-    public void save(Device device) {
-        repository.save(new DeviceEntity());
-    }
-}
