+++ modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/managed/ManagedLogEntity.java
@@ -5,7 +5,6 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
 import java.util.Objects;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -18,7 +17,7 @@ import lombok.experimental.SuperBuilder;
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ManagedLog extends BaseEntity {
+public class ManagedLogEntity extends BaseEntity {
 
     @EmbeddedId
     private LogId logId;
