+++ modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/member/MemberEntityJpaRepository.java
@@ -1,0 +1,7 @@
+package com.whoz_in.domain_jpa.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberEntityJpaRepository extends JpaRepository<MemberEntity, Long> {
+
+}
