+++ modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/member/MemberJpaRepository.java
@@ -1,7 +1,19 @@
 package com.whoz_in.domain_jpa.member;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+import com.whoz_in.domain.member.MemberRepository;
+import com.whoz_in.domain.member.model.Member;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
 
-public interface MemberJpaRepository extends JpaRepository<MemberEntity, Long> {
+@Repository
+@RequiredArgsConstructor
+public class MemberJpaRepository implements MemberRepository {
+  private final MemberConverter converter;
+  private final MemberEntityJpaRepository jpaRepository;
 
+  @Override
+  public void save(Member member) {
+    MemberEntity memberEntity = converter.from(member);
+    jpaRepository.save(memberEntity);
+  }
 }
